Content_Generation_System/requirement.txt
asgiref==3.8.0
aspose-words==24.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
django-admin-interface==0.28.6
django-colorfield==0.11.0
idna==3.6
Markdown==3.6
pillow==10.2.0
python-slugify==8.0.4
requests==2.31.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
